// src/app/enum/EPermission.enum.ts
export enum EPermission {
  ADD_USER = 'ADD_USER',
  UPDATE_USER = 'UPDATE_USER',
  ADMIN_UPDATE_USER = 'ADMIN_UPDATE_USER',
  FIND_USER = 'FIND_USER',
  GET_USER = 'GET_USER',
  DELETE_USER = 'DELETE_USER',
  LOCK_USER = 'LOCK_USER',
  UNLOCK_USER = 'UNLOCK_USER',

  DATA_RECOVERY = 'DATA_RECOVERY',

  ADD_ROLE = 'ADD_ROLE',
  UPDATE_ROLE = 'UPDATE_ROLE',
  FIND_ROLE = 'FIND_ROLE',
  GET_ROLE = 'GET_ROLE',
  DELETE_ROLE = 'DELETE_ROLE',

  ADD_HOUSE = 'ADD_HOUSE',
  UPDATE_HOUSE = 'UPDATE_HOUSE',
  FIND_HOUSE = 'FIND_HOUSE',
  GET_HOUSE = 'GET_HOUSE',
  DELETE_HOUSE = 'DELETE_HOUSE',
}